// packages/components/src/components/badge/badge.types.ts
export type BadgeVariant = 'count' | 'status' | 'label';
export type BadgeSize = 'sm' | 'md';
export type BadgeStatus = 'online' | 'busy' | 'away' | 'offline';
export type BadgeColor =
  | 'neutral'
  | 'mauve'
  | 'green'
  | 'red'
  | 'orange'
  | 'blue'
  | 'purple'
  | 'custom';
export type BadgeOverlayPosition = 'top-right' | 'top-left' | 'bottom-right' | 'bottom-left';